pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moodtape"
version = "0.1.0"
description = "Telegram bot that creates music playlists based on emotional descriptions"
authors = [
    {name = "Moodtape Team", email = "team@moodtape.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-telegram-bot>=20.0",
    "openai>=1.0.0",
    "spotipy>=2.22.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",  # for Apple Music JWT tokens
    "cryptography>=41.0.0",  # for Apple Music private key handling and JWT ES256 algorithm
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/moodtape/moodtape"
Repository = "https://github.com/moodtape/moodtape"
Issues = "https://github.com/moodtape/moodtape/issues"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
] 